Support additional ACME providers
By default, SSL certificates are provisioned using Let's Encrypt's
production directory. If the only configuration provided is
`SSL_DOMAIN`, that's the behaviour we want.

However, it can also be useful to specify an alternate provider, so
we'll expose a few extra environment variable to make that possible:

`ACME_DIRECTORY` can be used to specify a different directory to use.
For example, the Let's Encrypt staging environment, or another provider
entirely.

Some providers require EAB credentials in order to issue certificates.
To support these providers, we allow prodiving them in `EAB_KID` and
`EAB_HMAC_KEY`