Don't duplicate keys when replacing cache items
When setting a cache item that replaces an existing item, we weren't
accounting for the fact that `keys` already held the item's key, and
`size` already included the items size. This means that after a
replacement, both `keys` and `size` would be wrong.

In the worst case this could lead to a crash when attempting to evict
old items, if we choose a key to evict that no longer contains a
matching item.