Preserve existing X-Forwarded-* headers
For X-Forwarded-Host and X-Forwarded-Proto, we should only populate
these ourselves if they weren't present in the original request. If they
were present, we should preserve them.

Otherwise when running behind another proxy, we may lose information
that it has provided. For example, where a downstream proxy terminated
TLS and set `X-Forwarded-Proto` to `https`, and connects to us over
plain `http`, we would have wrongly reset `X-Forwarded-Proto` to `http`,
which will mislead the upstream.