Only forward X-Forwarded-* when not using TLS
In [8d19698d] we added forwarding of existing X-Forwarded-* headers,
because these are often needed when running behind another proxy layer.

However, we shouldn't be trusting the headers as set by the original
client. So in the case where there is no additional proxy layer between
us and the client, we should drop the headers instead.

Given that our TLS mode is intended for direct client access (and
requires direct access for certificate provisioning), let's make the
condition as follows:

- When running with TLS, assume requests are from external clients, and
  drop the headers.
- When running without TLS, allow that there may be another proxy
  terminating TLS downstream of us, and keep the headers.

In a case where the default logic does not apply, the `FORWARD_HEADERS`
env can be used to override it.