Ensure all response writers are flushable
When using custom response writers in middleware, we need to ensure they
all implement `http.Flusher`. Otherwise, the resulting handler will not
be flushable, which means the reverse proxy has no way to flush response
content to the client.

Not being able to flush content breaks response types like SSE, which
depend on sending chunked data as it happens, rather than buffering it
all to the end.